fix test-file and list.hpp

+++ vasyakin.kirill/common/list.hpp
@@ -8,7 +8,7 @@
 namespace vasyakin
 {
 
-  const size_t MAX = std::numeric_limits< size_t >::max();
+  const long long MAX = std::numeric_limits< long long >::max();
 
   template< class T >
   class List
